+++ src/js/discussion/actions/actions.js
@@ -9,131 +9,6 @@ let voteId = 0
 //get request section
 export const GET_DISCUSSIONS_REQUEST = 'GET_DISCUSSIONS_REQUEST';
 
-export const getDiscussionsSuccess = (discussions) =>{
-  return {
-    type: 'GET_DISCUSSIONS_SUCCESS',
-    discussions: discussions
-  }
-} //sends action that is picked up by getDiscussionsReducer
-
-export const getDiscussions = () => {
-  return (dispatch) => {
-    return axios.get('/discussions')
-      .then(response => {
-        dispatch(getDiscussionsSuccess(response.data))
-      })
-      .catch(error => {
-        console.log(error);
-      })
-  }
-}
-
-export const getCampsSuccess = (camps) => {
-  var contributed = false;
-  if(camps.discussionContribution && camps.discussionContribution.length > 0) {
-    contributed = true;
-  }
-  return {
-    type: 'GET_CAMPS_SUCCESS',
-    camps: camps.data,
-    contributed: contributed
-  }
-} //sends action that is picked up by getCampsReducer
-
-export const getCamps = (discussionId, fullname) => {
-  return (dispatch) => {
-    return axios.get(`/discussion/${discussionId}/${fullname}`)
-      .then(response => {
-        dispatch(getCampsSuccess(response.data))
-      })
-      .catch(error => {
-        console.log(error);
-      })
-  }
-}
-
-export const getCommentsSuccess = (comments) => {
-  return {
-    type: 'GET_COMMENTS_SUCCESS',
-    comments: comments.data
-  }
-} //sends action that is picked up by getCommentsReducer
-
-export const getComments = (campId) => {
-  return (dispatch) => {
-    return axios.get('/comments/' + campId)
-      .then(response => {
-        dispatch(getCommentsSuccess(response.data))
-      })
-      .catch(error => {
-        console.log(error);
-      })
-  }
-}
-
-//post request section
-
-export const createDiscussionSuccess = (discussion) => {
-  return {
-    type: 'CREATE_DISCUSSION_SUCCESS',
-    id: discussion.id,
-    input: discussion.input,
-    user_id: discussion.user_id
-  }
-} //sends action that's picked up by discussionReducer
-
-export const createCampSuccess = (camp) => {
-  return {
-    type: 'CREATE_CAMP_SUCCESS',
-    id: camp.id,
-    discussion_id: camp.discussion_id,
-    input: camp.input
-  }
-} //sends action that's picked up by campReducer
-
-export const createCampPost = (camp) => {
-  return (dispatch) => {
-    return axios.post('/commonground', camp)
-      .then(response => {
-        dispatch(createCampSuccess(response.data[0]))
-      })
-  }
-}
-
-export const createCommentSuccess = (comment) => {
-  console.log('created comment', comment)
-  return {
-    type: 'CREATE_COMMENT_SUCCESS',
-    id: comment.id,
-    commonground_id: comment.commonground_id,
-    input: comment.input,
-    user_id: comment.user_id,
-    fullname: comment.fullname,
-    facebookpicture: comment.facebookpicture,
-    delta: 0,
-    createdat: comment.createdat
-  }
-}
-
-export const votesSuccess = (vote) => {
-  return {
-    type: 'VOTE_SUCCESS',
-    commentId: vote.id,
-    delta: vote.delta,
-    upvotecounter: vote.upvotecounter,
-    downvotecounter: vote.downvotecounter
-  }
-}
-
-export function votesPost(vote) {
-  return (dispatch) => {
-    return axios.post('/vote', vote)
-      .then(response => {
-        dispatch(votesSuccess(response.data))
-      })
-  }
-}
-
 export const cachingFbData = (fbUser) => {
   console.log('cachingFBData', fbUser)
   console.log('!#!#!#!#!#!#!#!',fbUser.facebookpicture)
@@ -180,28 +55,5 @@ export function sendingFbDataSuccess(fbUser) {
     fbLocale: fbUser.locale,
     fbEmail: fbUser.email,
     fbPicture: fbUser.facebookpicture
-  }
-}
-
-export const contributedOnce = () => {
-  return {
-    type: 'CONTRIBUTED'
-  }
-}
-
-export const contributeAgainSuccess = () => {
-  return {
-    type: 'CONTRIBUTE_AGAIN'
-  }
-}
-
-export const contributeAgain = (currentContribution, callback) => {
-  return (dispatch) => {
-    return axios.post('/delete', currentContribution)
-      .then(response => {
-        dispatch(getCommentsSuccess(response))
-        dispatch(contributeAgainSuccess())
-        callback()
-      })
   }
 }
