+++ src/js/discussion/camps/AddCamp.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { connect } from 'react-redux'
-import { createCampPost } from '../actions/actions'
+import { createCampPost } from './campActions'
 import { InputGroup, Button, FormControl, HelpBlock, FormGroup, ControlLabel, Grid, Row, Col, Media } from 'react-bootstrap';
 require('./camp.css')
