+++ src/js/discussion/camps/CampParent.js
@@ -4,7 +4,6 @@ import AddCamp from './AddCamp'
 import CampList from './CampList'
 import { bindActionCreators } from 'redux'
 import { connect } from 'react-redux'
-import { getCamps } from '../actions/actions'
 import BackButton from './BackButton'
 import FaceBookIntegration from './FaceBookIntegration'
 import ProfileButton from './ProfileButton'
